auto_process.py 0.1.0: major update changing how fastq files are handled.
This commit affects the 'setup_analysis_dirs' and 'archive' commands; by default
fastqs are no longer symlinked but are hardlinked from project dirs, and so are
essentially copied upon archiving (use the --link-to-fastqs' option to revert to
the old behaviour). Also full fastq names are always used (use the
'--short-fastq-names' option to revert to the old behaviour).

If fastqs are not links then 'archive' will automatically exclude the unaligned
directory from being copied to the archive area.