qc/runqc: significant update to RunQC to set IlluminaQC per-project.
Significant update to the RunQC class which now forces the IlluminaQC
instance to be specified per-project at the point the project is added,
rather than a single instance for all projects at run-time.

This breaks backwards compatibility; the commit includes updates to
the 'run_qc' command, and to the run_qc.py and process_icell8.py
utilities, to accommodate these changes.