qc/pipeline: refactor 'RunIlluminaQC' task to improve efficiency for large datasets.
Refactors the 'RunIlluminaQC' task to run the 'illumina_qc.sh' script directly
(i.e. not via a function call), and split the version check into a separate
task which is executed only once. This should make the task startup faster and
less resource-hungry for datasets with large numbers of Fastqs.