qc/utils: reimplement 'set_cell_count_for_project'.
Reimplement the 'set_cell_count_for_project' function to remove code for determining the 10x software package outputs should be parsed for cell counts; instead the name of the appropriate package must be explicitly supplied as an argument.

Also updates the 'Set10xCellCount' task in the QC pipeline to accommodate the changes, and the invocations of the task within the pipeline.